Add containerize flag to avoid starting kubelet and collecting logs.

+++ test/e2e_node/services/services.go
@@ -43,6 +43,14 @@ import (
 // TODO(random-liu): Move this file to a separate package.
 var serverStartTimeout = flag.Duration("server-start-timeout", time.Second*120, "Time to wait for each server to become healthy.")
 
+// TODO(random-liu): Move kubelet start logic out of the test.
+// TODO(random-liu): Move log fetch logic out of the test.
+// There are different ways to start kubelet (systemd, initd, docker, rkt, manually started etc.)
+// and manage logs (journald, upstart etc.).
+// For different situation we need to mount different things into the container, run different commands.
+// It is hard and unnecessary to deal with the complexity inside the test suite.
+var containerize = flag.Bool("containerize", false, "If true, the test suite will not start kubelet, and fetch system log (kernel, docker, kubelet log etc.) to the report directory.")
+
 // E2EServices starts and stops e2e services in a separate process. The test uses it to start and
 // stop all e2e services.
 type E2EServices struct {
@@ -63,11 +71,13 @@ const servicesLogFile = "services.log"
 // services-mode to start e2e services in another process.
 func (e *E2EServices) Start() error {
 	var err error
-	// Create the manifest path for kubelet.
-	// TODO(random-liu): Remove related logic when we move kubelet starting logic out of the test.
-	framework.TestContext.ManifestPath, err = ioutil.TempDir("", "node-e2e-pod")
-	if err != nil {
-		return fmt.Errorf("failed to create static pod manifest directory: %v", err)
+	if !*containerize {
+		// Create the manifest path for kubelet.
+		// TODO(random-liu): Remove related logic when we move kubelet starting logic out of the test.
+		framework.TestContext.ManifestPath, err = ioutil.TempDir("", "node-e2e-pod")
+		if err != nil {
+			return fmt.Errorf("failed to create static pod manifest directory: %v", err)
+		}
 	}
 	testBin, err := osext.Executable()
 	if err != nil {
@@ -81,6 +91,7 @@ func (e *E2EServices) Start() error {
 		"--report-dir", framework.TestContext.ReportDir,
 		// TODO(random-liu): Remove the following flags after we move kubelet starting logic
 		// out of the test.
+		"--containerize="+strconv.FormatBool(*containerize),
 		"--node-name", framework.TestContext.NodeName,
 		"--disable-kubenet="+strconv.FormatBool(framework.TestContext.DisableKubenet),
 		// TODO: enable when flag is introduced in 1.5
@@ -97,12 +108,14 @@ func (e *E2EServices) Start() error {
 // Stop stops the e2e services.
 func (e *E2EServices) Stop() error {
 	defer func() {
-		// Cleanup the manifest path for kubelet.
-		manifestPath := framework.TestContext.ManifestPath
-		if manifestPath != "" {
-			err := os.RemoveAll(manifestPath)
-			if err != nil {
-				glog.Errorf("Failed to delete static pod manifest directory %s: %v", manifestPath, err)
+		if !*containerize {
+			// Cleanup the manifest path for kubelet.
+			manifestPath := framework.TestContext.ManifestPath
+			if manifestPath != "" {
+				err := os.RemoveAll(manifestPath)
+				if err != nil {
+					glog.Errorf("Failed to delete static pod manifest directory %s: %v", manifestPath, err)
+				}
 			}
 		}
 	}()
@@ -208,11 +221,13 @@ func (es *e2eService) start() error {
 		return err
 	}
 
-	s, err := es.startKubeletServer()
-	if err != nil {
-		return err
+	if !*containerize {
+		s, err := es.startKubeletServer()
+		if err != nil {
+			return err
+		}
+		es.services = append(es.services, s)
 	}
-	es.services = append(es.services, s)
 
 	err = es.startNamespaceController()
 	if err != nil {
@@ -281,7 +296,9 @@ func isJournaldAvailable() bool {
 
 func (es *e2eService) stop() {
 	glog.Info("Stopping e2e services...")
-	es.getLogFiles()
+	if !*containerize {
+		es.getLogFiles()
+	}
 	// TODO(random-liu): Use a loop to stop all services after introducing service interface.
 	// Stop namespace controller
 	if es.nsController != nil {
